requirements.txt.txt
python-telegram-bot==21.0.1
google-auth==2.28.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-core==2.15.0
google-api-python-client==2.117.0
